refactor: Rename MyTableCellRenderer to PlayerTableCellRenderer
